refactoring step time
