+++ src/traceml_ai/renderers/step_time/schema.py
@@ -1,55 +1,22 @@
-from __future__ import annotations
+"""Backward-compatible imports for Step Time domain contracts.
 
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, List, Optional
+New code should import these types from :mod:`traceml_ai.step_time.model`.
+This module remains temporarily so downstream integrations can migrate without
+an abrupt import-path break.
+"""
 
-if TYPE_CHECKING:
-    from traceml_ai.utils.step_time_window import StepTimeWindow
+from traceml_ai.step_time.model import (
+    StepCombinedTimeCoverage,
+    StepCombinedTimeMetric,
+    StepCombinedTimeResult,
+    StepCombinedTimeSeries,
+    StepCombinedTimeSummary,
+)
 
-
-@dataclass(frozen=True)
-class StepCombinedTimeSeries:
-    steps: List[int]
-    median: List[float]
-    worst: List[float]
-    sum: List[float]
-
-
-@dataclass(frozen=True)
-class StepCombinedTimeSummary:
-    window_size: int
-    steps_used: int
-    median_total: float
-    worst_total: float
-    worst_rank: Optional[int]
-    skew_ratio: Optional[float]
-    skew_pct: Optional[float]
-
-
-@dataclass(frozen=True)
-class StepCombinedTimeCoverage:
-    expected_steps: int
-    steps_used: int
-    completed_step: int
-    world_size: int
-    ranks_present: int
-    incomplete: bool
-
-
-@dataclass(frozen=True)
-class StepCombinedTimeMetric:
-    metric: str
-    clock: str  # "cpu" | "gpu" | "mixed"
-    series: Optional[StepCombinedTimeSeries]
-    summary: StepCombinedTimeSummary
-    coverage: StepCombinedTimeCoverage
-
-
-@dataclass(frozen=True)
-class StepCombinedTimeResult:
-    """Live Step Time state wrapping one canonical analyzed window."""
-
-    status_message: str = "OK"
-    window: Optional[StepTimeWindow] = None
-    training_strategy: str = "ddp"
-    had_ok: bool = False
+__all__ = [
+    "StepCombinedTimeCoverage",
+    "StepCombinedTimeMetric",
+    "StepCombinedTimeResult",
+    "StepCombinedTimeSeries",
+    "StepCombinedTimeSummary",
+]
